Fix variance (signed integer) overflow
In the variance calculations the difference is summed and later squared.
When the sum exceeds sqrt(2^31) the value is treated as a negative when
it is shifted which gives incorrect results.

To fix this we force the multiplication to be unsigned.

The alternative fix is to shift sum down by 4 before multiplying.
However that will reduce precision.

For 16x16 blocks the maximum sum is 65280 and sqrt(2^31) is 46340 (and
change).

This change is based on:
1698234 Missed some variance casts
fea3556 Fix variance overflow

+++ vp9/encoder/variance_c.c
@@ -37,7 +37,7 @@ unsigned int vp9_variance32x32_c(const unsigned char *src_ptr,
 
   variance(src_ptr, source_stride, ref_ptr, recon_stride, 32, 32, &var, &avg);
   *sse = var;
-  return (var - ((avg * avg) >> 10));
+  return (var - (((unsigned int)avg * avg) >> 10));
 }
 #endif
 
@@ -51,7 +51,7 @@ unsigned int vp9_variance16x16_c(const unsigned char *src_ptr,
 
   variance(src_ptr, source_stride, ref_ptr, recon_stride, 16, 16, &var, &avg);
   *sse = var;
-  return (var - ((avg * avg) >> 8));
+  return (var - (((unsigned int)avg * avg) >> 8));
 }
 
 unsigned int vp9_variance8x16_c(const unsigned char *src_ptr,
@@ -64,7 +64,7 @@ unsigned int vp9_variance8x16_c(const unsigned char *src_ptr,
 
   variance(src_ptr, source_stride, ref_ptr, recon_stride, 8, 16, &var, &avg);
   *sse = var;
-  return (var - ((avg * avg) >> 7));
+  return (var - (((unsigned int)avg * avg) >> 7));
 }
 
 unsigned int vp9_variance16x8_c(const unsigned char *src_ptr,
@@ -77,7 +77,7 @@ unsigned int vp9_variance16x8_c(const unsigned char *src_ptr,
 
   variance(src_ptr, source_stride, ref_ptr, recon_stride, 16, 8, &var, &avg);
   *sse = var;
-  return (var - ((avg * avg) >> 7));
+  return (var - (((unsigned int)avg * avg) >> 7));
 }
 
 
@@ -91,7 +91,7 @@ unsigned int vp9_variance8x8_c(const unsigned char *src_ptr,
 
   variance(src_ptr, source_stride, ref_ptr, recon_stride, 8, 8, &var, &avg);
   *sse = var;
-  return (var - ((avg * avg) >> 6));
+  return (var - (((unsigned int)avg * avg) >> 6));
 }
 
 unsigned int vp9_variance4x4_c(const unsigned char *src_ptr,
@@ -104,7 +104,7 @@ unsigned int vp9_variance4x4_c(const unsigned char *src_ptr,
 
   variance(src_ptr, source_stride, ref_ptr, recon_stride, 4, 4, &var, &avg);
   *sse = var;
-  return (var - ((avg * avg) >> 4));
+  return (var - (((unsigned int)avg * avg) >> 4));
 }
